refactor(daengle): 코드리뷰 토대로 리팩토링
