+++ apps/daengle/src/stores/oauth/index.tsx
@@ -1,0 +1,13 @@
+import { create } from 'zustand';
+
+interface AuthState {
+  accessToken: string | null;
+  setAccessToken: (token: string) => void;
+  clearAccessToken: () => void;
+}
+
+export const useAuthStore = create<AuthState>((set) => ({
+  accessToken: null,
+  setAccessToken: (token: string) => set({ accessToken: token }),
+  clearAccessToken: () => set({ accessToken: null }),
+}));
